requirements.txt
django                  == 3.2
gunicorn                == 20.0.4
flake8                  == 4.0
python-dotenv           == 0.21.1
psycopg2-binary         == 2.9 
django-debug-toolbar    == 3.4.0
django-split-settings   == 1.2.0